Add condition to not return error in case there are no configs to be removed

+++ utils/docker-utils.sh
@@ -137,8 +137,12 @@ docker::try_remove_volume() {
 docker::prune_configs() {
     local -r CONFIG_LABEL=${1:?"FATAL: remove_configs_by_label CONFIG_LABEL not provided"}
 
+    log info "Waiting for configs to be removed..."
     # shellcheck disable=SC2046
-    docker config rm $(docker config ls -qf label=name="$CONFIG_LABEL") &>/dev/null
+    if [[ -n $(docker config ls -qf label=name="$CONFIG_LABEL") ]]; then
+        docker config rm $(docker config ls -qf label=name="$CONFIG_LABEL") &>/dev/null
+    fi
+    overwrite "Waiting for configs to be removed... Done"
 }
 
 # Deploy a service, it will set config digests (in case a config is defined in the compose file)
